% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console_error.R
\name{console_error}
\alias{console_error}
\title{设置console中error的打印}
\usage{
console_error(text_color = "red", border_color = "red", bg_color = "white")
}
\value{
console
}
\description{
设置console中error的打印
}
